+++ src/Areas/Sql/SqlSetup.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using AzureMcp.Areas.Sql.Commands.Database;
+using AzureMcp.Areas.Sql.Commands.EntraAdmin;
 using AzureMcp.Areas.Sql.Services;
 using AzureMcp.Commands;
 using Microsoft.Extensions.DependencyInjection;
@@ -25,5 +26,13 @@ public class SqlSetup : IAreaSetup
         sql.AddSubGroup(database);
 
         database.AddCommand("show", new DatabaseShowCommand(loggerFactory.CreateLogger<DatabaseShowCommand>()));
+
+        var server = new CommandGroup("server", "SQL server operations");
+        sql.AddSubGroup(server);
+
+        var entraAdmin = new CommandGroup("entraadmin", "SQL server Microsoft Entra ID administrator operations");
+        server.AddSubGroup(entraAdmin);
+
+        entraAdmin.AddCommand("list", new EntraAdminListCommand(loggerFactory.CreateLogger<EntraAdminListCommand>()));
     }
 }